requirements.txt
streamlit==1.32.0
transformers==4.38.2
datasets==2.18.0
torch==2.2.1
langchain==0.1.11
langchain-community==0.0.27
psycopg2-binary==2.9.9
python-dotenv==1.0.1
sqlalchemy==2.0.27
flask==3.0.2
flask-socketio==5.3.6
pandas==2.2.1
numpy==1.26.4
scikit-learn==1.4.1
tqdm==4.66.2